refactor(server/auth): move protect() from authController to a new authMiddleware to separate concerns
