+++ server/controllers/authController.js
@@ -82,48 +82,6 @@ export const login = asyncHandler(async (req, res) => {
 });
 
 /**
- * Protect routes - middleware to check if user is authenticated
- */
-export const protect = asyncHandler(async (req, res, next) => {
-  // 1) Get token and check if it exists
-  let token;
-  if (
-    req.headers.authorization &&
-    req.headers.authorization.startsWith("Bearer")
-  ) {
-    token = req.headers.authorization.split(" ")[1];
-  }
-
-  if (!token) {
-    throw errorTypes.unauthorized(
-      "You are not logged in. Please log in to get access."
-    );
-  }
-
-  // 2) Verify token
-  const decoded = jwt.verify(token, process.env.JWT_SECRET);
-
-  // 3) Check if user still exists
-  const currentUser = await User.findById(decoded.id);
-  if (!currentUser) {
-    throw errorTypes.unauthorized(
-      "The user belonging to this token no longer exists."
-    );
-  }
-
-  // 4) Check if user changed password after the token was issued
-  if (currentUser.changedPasswordAfter(decoded.iat)) {
-    throw errorTypes.unauthorized(
-      "User recently changed password. Please log in again."
-    );
-  }
-
-  // GRANT ACCESS TO PROTECTED ROUTE
-  req.user = currentUser;
-  next();
-});
-
-/**
  * Restrict access to certain roles
  * @param  {...String} roles - Roles allowed to access the route
  */
